requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-debug-toolbar==5.2.0
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
dotenv==0.9.9
et_xmlfile==2.0.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
